removed unused files and correct import
