+++ index.js
@@ -1,9 +1,9 @@
 "use strict";
 
-const getConfig        = require('./readers/config_reader.js');
+const { getConfig }    = require('./readers/config_reader.js');
 const readSourceFiles  = require('./readers/source_reader.js');
 const createReaders    = require('./readers/create_readers.js');
-const parser           = require('./parser.js');
+const parser           = require('./parser/parser.js');
 
 getConfig()
   .then(readSourceFiles)
