+++ parser/parser.js
@@ -1,1 +1,11 @@
 "use strict";
+
+const parser = readers => {
+  readers.forEach((reader) => {
+    console.log("Printing Reader in Parser\n-----------------------\n");
+    console.log(reader.getTag());
+    console.log("\n\n");
+  });
+};
+
+module.exports = parser;
